chore: add auto import support in next
