+++ packages/next/build.js
@@ -1,0 +1,5 @@
+require('../../scripts/build')({
+  libraryDirectory: 'lib',
+  libraryName: '@alifd/next',
+  style: importPath => `${importPath}/style`
+})
